decode: fix integer warnings
Ticket: 4516

+++ src/datasets.c
@@ -141,7 +141,7 @@ static int ParseRepLine(const char *in, size_t ins, DataRepType *rep_out)
     }
 
     uint16_t v = 0;
-    int r = StringParseU16RangeCheck(&v, 10, strlen(ptrs[0]), ptrs[0], 0, USHRT_MAX);
+    int r = StringParseU16RangeCheck(&v, 10, (uint16_t)strlen(ptrs[0]), ptrs[0], 0, USHRT_MAX);
     if (r != (int)strlen(ptrs[0])) {
         SCLogError(SC_ERR_INVALID_NUMERIC_VALUE,
                 "'%s' is not a valid reputation value (0-65535)", ptrs[0]);
